Replace attr.is:value queries with attr == "value".
This came up as a result of discussions near the release of
taskwarrior-2.4.1.  It turns out that the 'attr.modifier:value' form of
querying is deprecated.  It won't go away for a long time, but we need
to eventually move to 'attr == "value"'.

This approach makes taskw perhaps more "helpful" than it should be in
that the query that the user pretty explicitly specifies to
python-taskw, gets changed under the covers before it is handed off to
taskwarrior.

This adds an extra test that failed before the code in
``taskw/utils.py`` was introduced.  That new code change makes the test
pass (and introduces no other new failures in the test suite).  Tox is
happy on a local run.