Pass description as positional argument.
Similarly to #134, passing description via description:{description}
doesn't appear to be documented and can lead to problems.

See ralphbean/bugwarrior#733 and consider the following scenario:

```sh
$ task --version
2.5.1
$ task add description:"depends filter does not work with ID"
The expression could not be evaluated.
$ task add "depends filter does not work with ID"
Created task 1.
```

I've tried variations of the previous with various quoting and shells
and have gotten the same result.