verbose must be first param for taskwarrior 2.5.2
There is a bug in version of taskwarrior that the rc.verbose=new-uuid
argument will only work if it is the first argument.

https://github.com/GothenburgBitFactory/taskwarrior/issues/1953

This change moves the 'verbose' config option to be first in
DEFAULT_CONFIG_OVERRIDES.

Without this change the test suite fails and it isn't possible to create
new tasks.